pyproject.toml
[project]
name = "weather-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.1",
    "openmeteo-requests>=1.7.5",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
]
